requirements.txt
bravado
bravado_core
jsonschema
pyramid
pytest
PyYAML
webtest
wsgiref
